Ticket #25: Open Watcom Support, Check for Borland C, special instructions added
git-svn-id: https://svn.code.sf.net/p/hwgui/code/trunk/hwgui@2849 66345640-f411-406e-acba-706ed0d45c9c

+++ source/winapi/control.c
@@ -853,10 +853,10 @@ HB_FUNC( HWG_TREEADDNODE )
 
    if( tvi.mask & TVIF_IMAGE )
       if ( tvi.iImage )
-         DeleteObject( ( HGDIOBJ ) tvi.iImage );
+         DeleteObject( ( HGDIOBJ ) ( UINT ) tvi.iImage );
    if( tvi.mask & TVIF_SELECTEDIMAGE )
       if ( tvi.iSelectedImage )
-         DeleteObject( ( HGDIOBJ ) tvi.iSelectedImage );
+         DeleteObject( ( HGDIOBJ ) ( UINT ) tvi.iSelectedImage );
 
    hb_strfree( hStr );
 }
@@ -2289,3 +2289,4 @@ HB_FUNC( HWG_DEFUSERLANG )
 }
 
 /* ====================== EOF of control.c ======================= */
+
